fix: preserve paged relation graph reads
